app/control/pyproject.toml
[project]
name = "aionos-control"
version = "0.1.0"
description = "AION-OS Control Plane (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.9",
  "httpx>=0.27",
  "redis>=5",
  "orjson>=3.10",
  "pydantic-settings>=2.2",
  "asyncpg>=0.29",
  "motor>=3.3",
  "qdrant-client>=1.7",
  "aiokafka>=0.10",
  "python-json-logger>=2.0",
  "python-dotenv>=1.0",
  "structlog>=24.1",
  "protobuf>=4.25",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "minio>=7.2",
  "clickhouse-connect>=0.7",
  "python-multipart>=0.0.9",
  "psycopg2-binary>=2.9"
]

[project.optional-dependencies]
ml-explain = ["shapash","aix360","cleanlab"]
ml-timeseries = ["stumpy","pysindy"]
ml-optim = ["nevergrad","cvxpy>=1.5","cvxpylayers"]
ml-automl = ["pycaret"]
ml-graphviz = ["pygwalker","hypertools"]
ml-features = ["featuretools"]
ml-ppl = ["pyro-ppl"]
ml-ui = ["taipy"]
notebook-addons = ["mitoinstaller","mitosheet","jupyterlab>=4"]
dev = ["pytest","pytest-asyncio","pytest-cov","ruff","mypy"]
all-ml = [
  "aix360","shapash","cleanlab",
  "stumpy","pysindy",
  "nevergrad","cvxpy","cvxpylayers",
  "pycaret","pygwalker","hypertools",
  "featuretools","pyro-ppl","taipy",
  "mitoinstaller","mitosheet"
]

[build-system]
requires = ["setuptools>=68","wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "aionos_control*"]
